Make const fields in `InjectedConstants` static readonly
Constants will be substituted at compile time in the used locations. That's not what we are looking for.
We want to change the value of those fields in the assembly, to affect all the code that used the fields.
Now we have to find the exact initialization instructions in the cctor.